4.21: auth for delete blog

+++ part4/blog-list/controllers/blogs.js
@@ -41,6 +41,17 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+
+  const decodedToken = jwt.verify(request.token, process.env.JWT_SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(request.params.id)
+  if (user.id.toString() !== blog.user.toString()) {
+    return response.status(401).json({ error: 'unauthorized' })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).send()
 })
